Trying to pass in arguments
I'm adding this one because I figured out why I can't reset the game. The code is using the global open_positions and pst variables instead of the ones within main(){}.
I tried to then pass in the main variables into each function but nothing gets returned when I enter a button to play the game. I need to fix that. Once that's done I can make the game restart once it's over and the project will be done.